listener/requirements.txt
TikTokLive>=5.0.0
websocket-client>=1.6.0
